refactor(init): switch from project init to create-config functionality
